+++ src/pages/dealershipDetails/components/DocList.js
@@ -3,11 +3,11 @@ import { useSnackbar } from 'notistack';
 import React, { useState } from 'react';
 import { useQuery, useQueryClient } from 'react-query';
 import DocListPreview from './DocListPreview';
-import FileUpload from '../../../components/FileUpload';
-import { URL } from '../../../config/serverUrls';
-import { getDealershipCheckList } from '../../../services/dealerships.service';
+import FileUpload, { FILE_FORMAT_ALL } from '../../../components/FileUpload';
 import { permissionCheck } from '../../../components/UserCan/UserCan';
+import { URL } from '../../../config/serverUrls';
 import { rulesList } from '../../../config/userRules';
+import { getDealershipCheckList } from '../../../services/dealerships.service';
 
 const DeleteButton = withStyles(() => ({
   root: {
@@ -119,7 +119,7 @@ const DocList = ({ id, currentUser }) => {
 
   return (
     <div className={classes.wrapper}>
-      {showUpload && <FileUpload handleSave={handleSave} id={id} data={rowData} title='Upload Dealership Document' open={showUpload} onCloseUploader={onCloseUploader} />}
+      {showUpload && <FileUpload handleSave={handleSave} id={id} data={rowData} title='Upload Dealership Document' open={showUpload} onCloseUploader={onCloseUploader} FILE_FORMAT={rowData.doc_id == '17' ? FILE_FORMAT_ALL : undefined} />}
       <Typography variant="h5" align={'Left'} className={classes.title}>
         Dealership Documents
       </Typography>