+++ src/pages/dealershipDetails/components/DocListPreview.js
@@ -2,6 +2,7 @@ import { Badge, Button, Typography, Dialog, DialogContent, DialogContentText, Co
 import { makeStyles } from '@material-ui/core/styles';
 import Tooltip from '@material-ui/core/Tooltip';
 import AddIcon from '@material-ui/icons/Add';
+import AudiotrackIcon from '@material-ui/icons/Audiotrack';
 import InfoCircleOutlined from '@material-ui/icons/InfoOutlined';
 import ListAltIcon from '@material-ui/icons/ListAlt';
 import PictureAsPdfIcon from '@material-ui/icons/PictureAsPdf';
@@ -15,6 +16,7 @@ import { deleteDocsImage } from '../../../services/dealerships.service';
 
 const imgFileTypes = ['jfif', 'pjpeg', 'jpeg', 'pjp', 'jpg', 'png'];
 const csvFileTypes = ['csv', 'xls', 'xlsx'];
+const audioFileTypes = ['mp3', 'wav', 'm4a']
 const useStyles = makeStyles((theme) => ({
   root: {
     padding: 10,
@@ -93,14 +95,16 @@ const DocPreview = ({ fileType, url, DocName, updatedDateTime, file_name, fileId
           <Tooltip title={`${file_name} (${updatedDateTime})`}>
             <span>
               <div className={classes.container}
-                onClick={() => csvFileTypes.includes(fileType) ? window.open(url) : setImageModal({ open: true, image: url, type: fileType })}
+                onClick={() => csvFileTypes.includes(fileType) ? window.open(url) : audioFileTypes.includes(fileType) ? window.open(url) : setImageModal({ open: true, image: url, type: fileType })}
               >
                 {
                   imgFileTypes.includes(fileType) ?
                     <img src={url} height="100%" width="100%" style={{ borderRadius: 6, padding: 1, objectFit: 'cover' }} alt={url} />
                     : fileType === 'pdf' ?
-                      <PictureAsPdfIcon style={{ color: '#63686E' }} />
-                      : <ListAltIcon style={{ color: '#63686E' }} />
+                      <PictureAsPdfIcon style={{ color: '#63686E' }} /> 
+                      : audioFileTypes.includes(fileType) ? 
+                        <AudiotrackIcon style={{ color: '#63686E' }} />
+                        : <ListAltIcon style={{ color: '#63686E'}} />
                 }
                 {
                   !editable &&