+++ src/pages/dealershipDetails/components/ScoreCard.js
@@ -1,4 +1,4 @@
-import { Typography, makeStyles, Button, Tabs, Tab, Box, CircularProgress, AppBar } from '@material-ui/core';
+import { Typography, makeStyles, Button, Tabs, Tab, Box, CircularProgress, AppBar, Backdrop, IconButton, Tooltip } from '@material-ui/core';
 import PublishIcon from '@material-ui/icons/Publish';
 import { useSnackbar } from 'notistack';
 import PropTypes from 'prop-types';
@@ -8,6 +8,7 @@ import { URL } from '../../../config/serverUrls';
 import { getScoreCard } from '../../../services/common.service';
 import BankingInputsTable from '../ScoreCardTables/BankingInputsTable';
 import BureauInputTable from '../ScoreCardTables/BureauInputTable';
+import GetAppIcon from '@material-ui/icons/GetApp';
 import CamInputTable from '../ScoreCardTables/CamInputTable';
 import CoappTable from '../ScoreCardTables/CoappTable';
 import DeviationsInputTable from '../ScoreCardTables/DeviationsInputTable';
@@ -16,6 +17,9 @@ import FixedObligationsTable from '../ScoreCardTables/FixedObligationsTable';
 import OmcSaleTable from '../ScoreCardTables/OmcSaleTable';
 import OtherInputsTable from '../ScoreCardTables/OtherInputsTable';
 import ScoreCardInputTable from '../ScoreCardTables/ScoreCardInputTable';
+import { head } from 'lodash';
+import { permissionCheck } from '../../../components/UserCan/UserCan';
+import { rulesList } from '../../../config/userRules';
 
 const useStyles = makeStyles(() => ({
   title: {
@@ -33,7 +37,10 @@ const useStyles = makeStyles(() => ({
   },
   buttonRipple: {
     borderRadius: '10px 10px 0px 0px',
-  }
+  },
+  backdrop: {zIndex: '2', position: 'absolute', margin: '-16px'},
+  backdropRoot: {display: 'flex', justifyContent: 'space-around', alignItems: 'center'},
+  caption: {color: 'rgb(0,0,0,0.3)'}
 }))
 
 function TabPanel(props) {
@@ -74,8 +81,11 @@ const ScoreCard = ({currentUser, dealership_id}) => {
   const queryClient = useQueryClient()
   const [tabValue, setTabValue] = useState(0)
   const [loading, setLoading] = useState(false)
+  const [notification, setNotification] = useState('Uploading score card')
   const { enqueueSnackbar } = useSnackbar();
   const { data: scoreCardData = [] } = useQuery('scorecard', () => getScoreCard(dealership_id), {refetchOnWindowFocus: false})
+  const metaData = head(scoreCardData?.metadata?.scorecard_meta)
+  const external = !permissionCheck(currentUser.role_name, rulesList.external_view)
 
   const scoreCardTabs = [
     'OMC Sale Data',
@@ -89,6 +99,19 @@ const ScoreCard = ({currentUser, dealership_id}) => {
     'Score Card'
   ]
 
+  const delay = async (ms = 1000) =>
+  new Promise(resolve => setTimeout(resolve, ms))
+
+  const loaderNotifications = [
+    'Analyzing OMC Sale Data...',
+    'Analyzing Bureau inputs...',
+    'Analyzing Banking inputs...',
+    'Analyzing ITR, P&L inputs...',
+    'Calculating Deviations...',
+    'Checking Eligibility...',
+    'Almost done!'
+  ]
+
   let cardData = scoreCardData?.co_app_sheet?.co_app_summary_data?.find(d => {
     d.type === 'co_app_1' && scoreCardTabs.push('Bureau-Ind Co-app 1')
     d.type === 'co_app_2' && scoreCardTabs.push('Bureau-Ind Co-app 2')
@@ -97,6 +120,12 @@ const ScoreCard = ({currentUser, dealership_id}) => {
 
   const handleChange = (event, newValue) => {
     setTabValue(newValue)
+  }
+
+  const handleDownload = () => {
+    if(metaData?.file_url){
+      window.open(metaData?.file_url, '_blank')
+    }
   }
 
   const onChangeHandler = (event) => {
@@ -116,18 +145,28 @@ const ScoreCard = ({currentUser, dealership_id}) => {
         return res.json()
       })
       .then(({status, message}) => {
-        setLoading(false)
         if(status === 'SUCCESS'){
-          queryClient.invalidateQueries('scorecard')
-          enqueueSnackbar(message, {
-            anchorOrigin: {
-              vertical: 'top',
-              horizontal: 'right',
-            },
-            variant: 'success'
-          })
+          const notify = async () => {
+            for (let i = 0; i < loaderNotifications.length; i++) {
+              setNotification(loaderNotifications[i])
+              if (i+1 == loaderNotifications?.length){
+                setLoading(false)
+                queryClient.invalidateQueries('scorecard')
+                enqueueSnackbar(message, {
+                  anchorOrigin: {
+                    vertical: 'top',
+                    horizontal: 'right',
+                  },
+                  variant: 'success'
+                })
+              }
+              await delay(1500)
+            }
+          }
+          notify()
         }
         else {
+          setLoading(false)
           enqueueSnackbar(message, {
             anchorOrigin: {
               vertical: 'top',
@@ -151,25 +190,48 @@ const ScoreCard = ({currentUser, dealership_id}) => {
   };
 
   return (
-    <>
+    <div style={{position: 'relative'}}>
       <div className={classes.title}>
-        <Typography variant="h5">Eligibility Score Card</Typography>
-        <Button
-          variant='outlined'
-          color='primary'
-          name='csv'
-          id='file'
-          component="label"
-          disabled={loading}
-          startIcon={loading ? <CircularProgress size={14} /> : <PublishIcon fontSize='small' />}
-        >Upload score card
-          <input 
-            type="file"
-            hidden
-            accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-            onChange={onChangeHandler}
-          />
-        </Button>
+        <div style={{display: 'flex', flexDirection: 'column'}}>
+          <Typography variant="h5">Eligibility Score Card</Typography>
+          {
+            metaData?.uploaded_date &&
+            <Typography variant="caption" className={classes.caption}>Uploaded On: {metaData?.uploaded_date || '-'}</Typography>
+          }
+          {
+            metaData?.uploaded_by &&
+            <Typography variant="caption" className={classes.caption}>Uploaded By: {metaData?.uploaded_by || '-'}</Typography>
+          }
+        </div>
+        <div>
+          {
+            metaData?.file_url &&
+            <Tooltip title="Download Score Card">
+              <IconButton size="small" style={{marginRight: 12}} onClick={handleDownload}>
+                <GetAppIcon />
+              </IconButton>
+            </Tooltip>
+          }
+          {
+            external &&
+            <Button
+              variant='outlined'
+              color='primary'
+              name='csv'
+              id='file'
+              component="label"
+              disabled={loading}
+              startIcon={loading ? <CircularProgress size={14} /> : <PublishIcon fontSize='small' />}
+            >Upload score card
+              <input
+                type="file"
+                hidden
+                accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
+                onChange={onChangeHandler}
+              />
+            </Button>
+          }
+        </div>
       </div>
       <div className={classes.tabsRoot}>
         <AppBar position='static' color='default' style={{zIndex: '0'}}>
@@ -227,7 +289,13 @@ const ScoreCard = ({currentUser, dealership_id}) => {
           <CoappTable data={scoreCardData?.co_app_sheet} type='co_app_3' />
         </TabPanel>
       </div>
-    </>
+      <Backdrop open={loading} className={classes.backdrop} >
+        <div className={classes.backdropRoot}>
+          <CircularProgress style={{color: 'white'}} size={25} />
+          <Typography variant='body1' style={{color: 'white', marginLeft: 16}}>{notification}</Typography>
+        </div>
+      </Backdrop>
+    </div>
   )
 }